Remove noscroll feature
This commit includes the removal of the noscroll feature from the application. The changes include modifications to the asset-manifest.json and index.html files, renaming of certain CSS and JS files, and modifications to the App.js file. Some files have been deleted and new ones have been added as part of this change.